+++ vnext/src-win/Libraries/Components/Pressable/Pressable.windows.js
@@ -9,9 +9,9 @@
  */
 
 import type {
-  LayoutEvent,
+  LayoutChangeEvent,
   MouseEvent,
-  PressEvent,
+  GestureResponderEvent,
   // [Windows
   BlurEvent,
   FocusEvent,
@@ -134,7 +134,7 @@ type Props = $ReadOnly<{
   /**
    * Called when this view's layout changes.
    */
-  onLayout?: ?(event: LayoutEvent) => mixed,
+  onLayout?: ?(event: LayoutChangeEvent) => mixed,
 
   /**
    * Called when the hover is activated to provide visual feedback.
@@ -149,22 +149,22 @@ type Props = $ReadOnly<{
   /**
    * Called when a long-tap gesture is detected.
    */
-  onLongPress?: ?(event: PressEvent) => mixed,
+  onLongPress?: ?(event: GestureResponderEvent) => mixed,
 
   /**
    * Called when a single tap gesture is detected.
    */
-  onPress?: ?(event: PressEvent) => mixed,
+  onPress?: ?(event: GestureResponderEvent) => mixed,
 
   /**
    * Called when a touch is engaged before `onPress`.
    */
-  onPressIn?: ?(event: PressEvent) => mixed,
+  onPressIn?: ?(event: GestureResponderEvent) => mixed,
 
   /**
    * Called when a touch is released before `onPress`.
    */
-  onPressOut?: ?(event: PressEvent) => mixed,
+  onPressOut?: ?(event: GestureResponderEvent) => mixed,
 
   /**
    * Called after the element loses focus.
@@ -339,7 +339,7 @@ function Pressable(
     accessibilityLiveRegion,
     accessibilityLabel,
     accessibilityState: _accessibilityState,
-    disabled: disabled == true,
+    disabled: disabled === true,
     focusable: focusable !== false,
     accessibilityValue,
     hitSlop,
@@ -360,7 +360,7 @@ function Pressable(
       onHoverOut,
       onLongPress,
       onPress,
-      onPressIn(event: PressEvent): void {
+      onPressIn(event: GestureResponderEvent): void {
         if (android_rippleConfig != null) {
           android_rippleConfig.onPressIn(event);
         }
@@ -370,7 +370,7 @@ function Pressable(
         }
       },
       onPressMove: android_rippleConfig?.onPressMove,
-      onPressOut(event: PressEvent): void {
+      onPressOut(event: GestureResponderEvent): void {
         if (android_rippleConfig != null) {
           android_rippleConfig.onPressOut(event);
         }