+++ vnext/src-win/Libraries/Components/ScrollView/ScrollViewNativeComponent.windows.js
@@ -42,7 +42,7 @@ export const __INTERNAL_VIEW_CONFIG: PartialViewConfig =
         },
         validAttributes: {
           contentOffset: {
-            diff: require('../../Utilities/differ/pointsDiffer'),
+            diff: require('../../Utilities/differ/pointsDiffer').default,
           },
           decelerationRate: true,
           disableIntervalMomentum: true,
@@ -126,10 +126,10 @@ export const __INTERNAL_VIEW_CONFIG: PartialViewConfig =
           canCancelContentTouches: true,
           centerContent: true,
           contentInset: {
-            diff: require('../../Utilities/differ/insetsDiffer'),
+            diff: require('../../Utilities/differ/insetsDiffer').default,
           },
           contentOffset: {
-            diff: require('../../Utilities/differ/pointsDiffer'),
+            diff: require('../../Utilities/differ/pointsDiffer').default,
           },
           contentInsetAdjustmentBehavior: true,
           decelerationRate: true,
@@ -148,7 +148,7 @@ export const __INTERNAL_VIEW_CONFIG: PartialViewConfig =
           scrollEnabled: true,
           scrollEventThrottle: true,
           scrollIndicatorInsets: {
-            diff: require('../../Utilities/differ/insetsDiffer'),
+            diff: require('../../Utilities/differ/insetsDiffer').default,
           },
           scrollToOverflowEnabled: true,
           scrollsToTop: true,
@@ -159,6 +159,9 @@ export const __INTERNAL_VIEW_CONFIG: PartialViewConfig =
           snapToInterval: true,
           snapToOffsets: true,
           snapToStart: true,
+          verticalScrollIndicatorInsets: {
+            diff: require('../../Utilities/differ/insetsDiffer').default,
+          },
           zoomScale: true,
           ...ConditionallyIgnoredEventHandlers({
             onScrollBeginDrag: true,