+++ vnext/src-win/Libraries/Components/View/View.windows.js
@@ -288,9 +288,9 @@ const View: component(
         // https://github.com/facebook/react-native/commit/66601e755fcad10698e61d20878d52194ad0e90c
         // But since Views are not currently supported in Text, we do not need the extra provider
         <TextAncestor.Consumer>
-          {hasTextAncestor => {
+          {consumerHasTextAncestor => {
             invariant(
-              !hasTextAncestor,
+              !consumerHasTextAncestor,
               'Nesting of <View> within <Text> is not currently supported.',
             );
             return (