+++ packages/@react-native-windows/automation-channel/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @react-native-windows/automation-channel
 
-<!-- This log was last generated on Thu, 15 May 2025 05:23:40 GMT and should not be manually modified. -->
+<!-- This log was last generated on Tue, 20 May 2025 06:25:35 GMT and should not be manually modified. -->
 
 <!-- Start content -->
+
+## 0.12.298
+
+Tue, 20 May 2025 06:25:35 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.958
+
+## 0.12.297
+
+Sat, 17 May 2025 05:25:32 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.957
+
+## 0.12.296
+
+Fri, 16 May 2025 05:27:55 GMT
+
+### Patches
+
+- Bump react-native-windows to v0.0.0-canary.956
 
 ## 0.12.295